Factor room generation into its own file
