+++ grid_maze_random_rooms.h
@@ -1,0 +1,3 @@
+#include "maze.h"
+
+int generate_rooms(int maze_height, int maze_width, int rooms_length, maze_room_t **rooms);
